DA-3703: Handle skipped schema/object check correctly
Previously, if skip criteria was provided as "sch.ojb"
and also "dropExtras" was specified then extra objects were not getting
dropped in the "sch" schema (even if they didn't match "obj").
Instead we want all extra objects that don't match "obj" in that schema
to be dropped. This commit does that.